Website Update
Completed steps 1 - 10 by creating an event listener to run an anonymous function when the page loads that loops through every item in the changingCells object collection, and for each item adds an onchange event handler that runs calcExp() function, and adds an event handler for the click event that runs the validateSummary() function when the button with the ID of "submitButton" is clicked. Also created the validateSummary() function that validates the data entry in the summary field, created the calcClass() function that sums up all of the data values for elements of the sumClass class, and created the calcExp() function that calculates the travel expenses from all categories and dates.